my-react-app/requirements.txt
fastapi
uvicorn
fitz
pillow
httpx